Make the Circle bounce.

+++ capstone-java/src/org/capstone/game/physics/RectPhysicsComponent.java
@@ -12,7 +12,7 @@ public class RectPhysicsComponent extends PhysicsComponent {
 	}
 
 	@Override
-	public void update(double elapsedTime) {
+	public void update(long elapsedTime) {
 		super.update(elapsedTime);
 	}
 }
